move scale algorithm to scale
