+++ game/magic/scale/scale.go
@@ -10,6 +10,28 @@ import (
 var ScaleAmount = 3.0
 var ScaledGeom ebiten.GeoM
 
+var ScreenScaleAlgorithm = ScaleAlgorithmNormal
+
+type ScaleAlgorithm int
+
+const (
+    // the scale2x
+    // https://www.scale2x.it/
+    ScaleAlgorithmScale ScaleAlgorithm = iota
+    ScaleAlgorithmXbr
+    ScaleAlgorithmNormal
+)
+
+func (algorithm ScaleAlgorithm) String() string {
+    switch algorithm {
+        case ScaleAlgorithmScale: return "scale"
+        case ScaleAlgorithmXbr: return "xbr"
+        case ScaleAlgorithmNormal: return "normal"
+    }
+
+    return ""
+}
+
 type UnscaledGeoM ebiten.GeoM
 
 func (unscaled *UnscaledGeoM) Scaled() ebiten.GeoM {
