+++ game/magic/util/image-cache.go
@@ -9,9 +9,8 @@ import (
     "image/color"
 
     "github.com/kazzmir/master-of-magic/lib/lbx"
-    "github.com/kazzmir/master-of-magic/lib/xbr"
+    // "github.com/kazzmir/master-of-magic/lib/xbr"
     "github.com/kazzmir/master-of-magic/game/magic/shaders"
-    "github.com/kazzmir/master-of-magic/game/magic/data"
     "github.com/hajimehoshi/ebiten/v2"
 )
 
@@ -29,8 +28,10 @@ type ImageCache struct {
 
     ShaderCache map[shaders.Shader]*ebiten.Shader
 
+    /*
     Scaler data.ScaleAlgorithm
     ScaleAmount int
+    */
 }
 
 func MakeImageCache(lbxCache *lbx.LbxCache) ImageCache {
@@ -38,8 +39,10 @@ func MakeImageCache(lbxCache *lbx.LbxCache) ImageCache {
         LbxCache: lbxCache,
         Cache:    make(map[string][]*ebiten.Image),
         ShaderCache: make(map[shaders.Shader]*ebiten.Shader),
+        /*
         Scaler: data.ScreenScaleAlgorithm,
         ScaleAmount: 1,
+        */
         // ScaleAmount: data.ScreenScale,
     }
 }
@@ -318,6 +321,7 @@ func Scale3x(input image.Image, smooth bool) image.Image {
 }
 
 func (cache *ImageCache) ApplyScale(input image.Image) image.Image {
+    /*
     if cache.ScaleAmount == 1 {
         return input
     }
@@ -339,6 +343,7 @@ func (cache *ImageCache) ApplyScale(input image.Image) image.Image {
             }
         case data.ScaleAlgorithmXbr: return xbr.ScaleImage(input, cache.ScaleAmount)
     }
+    */
 
     return input
 }
