+++ test/scale/main.go
@@ -159,7 +159,7 @@ func NewEngine(scenario int) (*Engine, error) {
     }, nil
 }
 
-func (engine *Engine) ChangeScale(scaleAmount int, algorithm data.ScaleAlgorithm) {
+func (engine *Engine) ChangeScale(scaleAmount int, algorithm scale.ScaleAlgorithm) {
     /*
     data.ScreenScale = 1
     data.ScreenScaleAlgorithm = algorithm
@@ -182,16 +182,16 @@ func (engine *Engine) Update() error {
     for _, key := range keys {
         switch key {
             case ebiten.KeyEscape, ebiten.KeyCapsLock: return ebiten.Termination
-            case ebiten.KeyF1: engine.ChangeScale(1, data.ScaleAlgorithmNormal)
-            case ebiten.KeyF2: engine.ChangeScale(2, data.ScaleAlgorithmNormal)
-            case ebiten.KeyF3: engine.ChangeScale(3, data.ScaleAlgorithmNormal)
-            case ebiten.KeyF4: engine.ChangeScale(4, data.ScaleAlgorithmNormal)
-            case ebiten.KeyF5: engine.ChangeScale(2, data.ScaleAlgorithmScale)
-            case ebiten.KeyF6: engine.ChangeScale(3, data.ScaleAlgorithmScale)
-            case ebiten.KeyF7: engine.ChangeScale(4, data.ScaleAlgorithmScale)
-            case ebiten.KeyF8: engine.ChangeScale(2, data.ScaleAlgorithmXbr)
-            case ebiten.KeyF9: engine.ChangeScale(3, data.ScaleAlgorithmXbr)
-            case ebiten.KeyF10: engine.ChangeScale(4, data.ScaleAlgorithmXbr)
+            case ebiten.KeyF1: engine.ChangeScale(1, scale.ScaleAlgorithmNormal)
+            case ebiten.KeyF2: engine.ChangeScale(2, scale.ScaleAlgorithmNormal)
+            case ebiten.KeyF3: engine.ChangeScale(3, scale.ScaleAlgorithmNormal)
+            case ebiten.KeyF4: engine.ChangeScale(4, scale.ScaleAlgorithmNormal)
+            case ebiten.KeyF5: engine.ChangeScale(2, scale.ScaleAlgorithmScale)
+            case ebiten.KeyF6: engine.ChangeScale(3, scale.ScaleAlgorithmScale)
+            case ebiten.KeyF7: engine.ChangeScale(4, scale.ScaleAlgorithmScale)
+            case ebiten.KeyF8: engine.ChangeScale(2, scale.ScaleAlgorithmXbr)
+            case ebiten.KeyF9: engine.ChangeScale(3, scale.ScaleAlgorithmXbr)
+            case ebiten.KeyF10: engine.ChangeScale(4, scale.ScaleAlgorithmXbr)
         }
     }
