+++ util/combat-simulator/main.go
@@ -678,7 +678,6 @@ func (engine *Engine) MakeBugUI() *ebitenui.UI {
 
 func (engine *Engine) MakeUI() *ebitenui.UI {
     imageCache := util.MakeImageCache(engine.Cache)
-    imageCache.ScaleAmount = 1
 
     face, _ := loadFont(19)
